+++ SadnaSrc/SadnaSrc/StoreCenter/OutsideModuleService.cs
@@ -17,7 +17,7 @@ namespace SadnaSrc.StoreCenter
         double CalculateItemPriceWithDiscount(string storeName, string productName, string _DiscountCode, int _quantity);
 
         bool HasActiveLottery(string storeName, string productName, double priceWantToPay);
-        void updateLottery(string storeName, string ProductName, int moenyPayed, string UserName);
+        void updateLottery(string storeName, string ProductName, double moenyPayed, string UserName);
 
     }
 }
